+++ packages/be-ts-db-lib/src/index.ts
@@ -18,6 +18,7 @@ import WorkoutMuscleGroupRepository from './repositories/workout/WorkoutMuscleGr
 import WorkoutSessionExerciseRepository from './repositories/workout/WorkoutSessionExerciseRepository.js';
 import WorkoutSessionRepository from './repositories/workout/WorkoutSessionRepository.js';
 import WorkoutSetRepository from './repositories/workout/WorkoutSetRepository.js';
+import GoogleAuthService from './services/GoogleAuthService.js';
 import DbOperationMetaData from './util/DbOperationMetaData.js';
 import DocumentDb from './util/DocumentDb.js';
 
@@ -33,6 +34,7 @@ export {
   DashboardUserConfigRepository,
   DbOperationMetaData,
   DocumentDb,
+  GoogleAuthService,
   UserRepository,
   WorkoutBaseRepository,
   WorkoutBaseWithUserIdRepository,