+++ packages/be-ts-lib/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [3.1.5] (2026-03-15)
+
+### ✅ Added
+
+- Added `jwtAccessSecret` field to the `Config` interface for signing JWT access tokens.
+
 ## 🔖 [3.1.4] (2026-03-13)
 
 ### 🏗️ Changed
@@ -260,6 +266,7 @@ No direct code changes; version bump for compatibility with new major versions o
 
 <!-- Link References -->
 
+[3.1.5]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.1.4...be-ts-lib-v3.1.5
 [3.1.4]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.1.3...be-ts-lib-v3.1.4
 [3.1.3]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.1.2...be-ts-lib-v3.1.3
 [3.1.2]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.1.1...be-ts-lib-v3.1.2