+++ packages/local-npm-registry/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [0.2.30] (2026-03-15)
+
+### 🏗️ Changed
+
+- Updated `@aneuhold/core-ts-lib` dependency to `^2.4.3`.
+
 ## 🔖 [0.2.29] (2026-03-13)
 
 ### 🏗️ Changed
@@ -183,6 +189,7 @@ No direct code changes; version bump for compatibility.
 
 <!-- Link References -->
 
+[0.2.30]: https://github.com/aneuhold/ts-libs/compare/local-npm-registry-v0.2.29...local-npm-registry-v0.2.30
 [0.2.29]: https://github.com/aneuhold/ts-libs/compare/local-npm-registry-v0.2.28...local-npm-registry-v0.2.29
 [0.2.28]: https://github.com/aneuhold/ts-libs/compare/local-npm-registry-v0.2.27...local-npm-registry-v0.2.28
 [0.2.27]: https://github.com/aneuhold/ts-libs/compare/local-npm-registry-v0.2.26...local-npm-registry-v0.2.27